+++ packages/cli/src/index.ts
@@ -1,19 +1,12 @@
-import './alias';
-import { defineCommand, CommandConfig } from './core/command'; // 会将 defineCommand 定义到 当前模块的exports对象上面
-import CommandInit from '@em-cli/em-plugin-init'; // 插件中因为以来 @em-cli/cli 这个包 如果先导入插件就会导致 内部 defineCommand 为 undefined
+import ECli, { CommandAddCmd, CommandAdminConfig } from '@em-cli/core';
+import CommandInit from '@em-cli/em-plugin-init';
 import CommandDev from '@em-cli/em-plugin-dev';
 import CommandBuild from '@em-cli/em-plugin-build';
 import CommandDeploy from '@em-cli/em-plugin-deploy';
 import CommandCreate from '@em-cli/em-plugin-create';
 import CommandTemplate from '@em-cli/em-plugin-template';
-import {
-  CommandAddCmd,
-  CommandAdminConfig,
-  CONFIG_PATH,
-  getConfigKey,
-  setConfigKey,
-} from '@/command';
-import ECli from '@/core';
+import CommandRenderer from '@em-cli/em-plugin-renderer';
+
 export function run(isDev: boolean) {
   const eCli = new ECli();
   eCli
@@ -23,16 +16,10 @@ export function run(isDev: boolean) {
     .addCommand(CommandDeploy)
     .addCommand(CommandCreate)
     .addCommand(CommandAdminConfig)
-    .addCommand(CommandTemplate);
+    .addCommand(CommandTemplate)
+    .addCommand(CommandRenderer);
   if (isDev) {
     eCli.addCommand(CommandAddCmd);
   }
   eCli.run();
 }
-export {
-  defineCommand,
-  CommandConfig,
-  CONFIG_PATH,
-  getConfigKey,
-  setConfigKey,
-};