+++ packages/em-plugin-init/src/index.ts
@@ -1,4 +1,4 @@
-import { defineCommand } from '@em-cli/em-cli';
+import { defineCommand } from '@em-cli/core';
 import loadFeature from './features/loadFeature';
 import { Creator } from './creator';