+++ packages/em-plugin-renderer/__tests__/index.test.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const emPluginDev = require('..');
+
+describe('@em-cli/em-plugin-em-plugin-renderer', () => {
+  it('needs tests');
+});