+++ packages/em-plugin-deploy/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.3](https://github.com/edisonLzy/em-cli/compare/@em-cli/em-plugin-deploy@0.2.2...@em-cli/em-plugin-deploy@0.2.3) (2022-03-22)
+
+**Note:** Version bump only for package @em-cli/em-plugin-deploy
+
 ## [0.2.2](https://github.com/edisonLzy/mono-cli/compare/@em-cli/em-plugin-deploy@0.2.1...@em-cli/em-plugin-deploy@0.2.2) (2021-11-08)
 
 **Note:** Version bump only for package @em-cli/em-plugin-deploy