+++ packages/em-plugin-dev/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/edisonLzy/em-cli/compare/@em-cli/em-plugin-dev@0.3.3...@em-cli/em-plugin-dev@0.4.0) (2022-03-22)
+
+### Features
+
+- **dev:** change port ([7833616](https://github.com/edisonLzy/em-cli/commit/78336162e417f1505ff7a7b5d929f1c034119c19))
+
+## [0.3.3](https://github.com/edisonLzy/em-cli/compare/@em-cli/em-plugin-dev@0.3.2...@em-cli/em-plugin-dev@0.3.3) (2022-03-22)
+
+**Note:** Version bump only for package @em-cli/em-plugin-dev
+
 ## [0.3.2](https://github.com/edisonLzy/mono-cli/compare/@em-cli/em-plugin-dev@0.3.1...@em-cli/em-plugin-dev@0.3.2) (2021-11-08)
 
 **Note:** Version bump only for package @em-cli/em-plugin-dev